feat(starter): add ability to configure http logging level
Adds the ability to configure the logging level for HTTP requests
and responses. This allows for more fine-grained control over the
amount of logging that is performed.

The following levels are supported:

- FATAL
- ERROR
- WARN
- INFO

The default level is INFO.

This feature is enabled by default. To disable it, set the
`http.logging.enabled` property to `false`.